==> exp_IMR_multitaskLearner_multi_surface/script_2_data_process_cathub.ipynb <==
# git clone https://github.com/Open-Catalyst-Project/ocp.git
# cd ocp
# pip install -e .

# pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-2.0.0+cu118.html
# pip install lmdb ase pymatgen
# pip install wandb tensorboard

# git clone https://github.com/SUNCAT-Center/CatKit.git
# cd CatKit
# SKLEARN_ALLOW_DEPRECATED_SKLEARN_PACKAGE_INSTALL=True
# pip install -e .

# # https://nbviewer.org/gist/mhoffman/c418acb6b3f928eb4ef71f4001d120d9

# import requests
# import pprint
# import sys
# import string
# import json
# import io
# import copy
# import ase.io
# import ase.calculators.singlepoint
# import pickle

# GRAPHQL = 'http://api.catalysis-hub.org/graphql'

# def fetch(query):
#     return requests.get(
#         GRAPHQL, {'query': query}
#     ).json()['data']

# def reactions_from_dataset(pub_id, page_size=10):
#     reactions = []
#     has_next_page = True
#     start_cursor = ''
#     page = 0
#     while has_next_page:
#         data = fetch("""{{
#       reactions(pubId: "{pub_id}", first: {page_size}, after: "{start_cursor}") {{
#         totalCount
#         pageInfo {{
#           hasNextPage
#           hasPreviousPage
#           startCursor
#           endCursor 
#         }}  
#         edges {{
#           node {{
#             chemicalComposition
#             surfaceComposition
#             facet
#             sites
#             coverages
#             reactants
#             products
#             Equation
#             reactionEnergy
#             activationEnergy
#             dftCode
#             dftFunctional
#             username
#             pubId
#             reactionSystems {{
#               name
#               systems {{
#                 energy
#                 InputFile(format: "json")
#               }}
#             }}  
#           }}  
#         }}  
#       }}    
#     }}""".format(start_cursor=start_cursor,
#                  page_size=page_size,
#                  pub_id=pub_id,
#                 ))
#         has_next_page = data['reactions']['pageInfo']['hasNextPage']
#         start_cursor = data['reactions']['pageInfo']['endCursor']
#         page += 1
#         print(has_next_page, start_cursor, page_size * page, data['reactions']['totalCount'])
#         reactions.extend(map(lambda x: x['node'], data['reactions']['edges']))

#     return reactions

# def aseify_reactions(reactions):
#     for i, reaction in enumerate(reactions):
#         for j, _ in enumerate(reactions[i]['reactionSystems']):
#             with io.StringIO() as tmp_file:
#                 system = reactions[i]['reactionSystems'][j].pop('systems')
#                 tmp_file.write(system.pop('InputFile'))
#                 tmp_file.seek(0)
#                 atoms = ase.io.read(tmp_file, format='json')
#             calculator = ase.calculators.singlepoint.SinglePointCalculator(
#                 atoms,
#                 energy=system.pop('energy')
#             )
#             atoms.set_calculator(calculator)
#             #print(atoms.get_potential_energy())
#             reactions[i]['reactionSystems'][j]['atoms'] = atoms
#         # flatten list further into {name: atoms, ...} dictionary
#         reactions[i]['reactionSystems'] = {x['name']: x['atoms']
#                                           for x in reactions[i]['reactionSystems']}
        
# if __name__=='__main__':
#     raw_reactions = reactions_from_dataset("MamunHighT2019") # this will get all 88587 reactions
#     # raw_reactions = reactions_from_dataset("FesterEdge2017")
    
#     reactions = copy.deepcopy(raw_reactions)
#     aseify_reactions(reactions)
    
#     # Save reactions to a file
#     with open('/curdir/datasets/reactions.pickle', 'wb') as f:
#         pickle.dump(reactions, f)
    
#     # # Load reactions from a file
#     # with open('/curdir/datasets/reactions.pickle', 'rb') as f:
#     #     loaded_reactions = pickle.load(f)
    

# !pip install ray

import pickle
import os
import numpy as np
import ase.io
from ase.constraints import FixAtoms
from ase.build import add_adsorbate, molecule, surface
from pymatgen.ext.matproj import MPRester
from pymatgen.core.surface import generate_all_slabs, SlabGenerator
from pymatgen.io.ase import AseAtomsAdaptor
from ocpmodels.common.relaxation.ase_utils import OCPCalculator
import json
import ray
import pandas as pd
from transformers import AutoTokenizer, AutoModel, pipeline

with open('datasets/reactions.pickle', 'rb') as f:
    loaded_reactions = pickle.load(f)

len(loaded_reactions)

loaded_reactions[4000:4001]

# m = MPRester('Yct0KDbJbqMLWluZEovkwrLXh2VRHXbc')

ray.init(ignore_reinit_error=True)

config_yml_path = "ocp/configs/is2re/all/dimenet_plus_plus/dpp.yml"
checkpoint_path = "dimenetpp_all.pt"

# Define the calculator
calc_id = ray.put(OCPCalculator(config_yml=config_yml_path, checkpoint=checkpoint_path))

def get_energy(ase_slab, ads_symbol):
    
    adslab = ase_slab.copy()
    adslab.calc = ray.get(calc_id)
    # print(adslab)
    
    slab_energy = adslab.get_potential_energy()
    
    # add the molecule
    adsorbate = molecule(ads_symbol)
    add_adsorbate(adslab, adsorbate, 3, offset=(1, 1))
    # print(adslab)
    
    ads_energy = adslab.get_potential_energy()
    
    # because this is a fingerprint we do not need to reference to gas molecules 
    # because this is the same constant across surfaces
    
    return ads_energy - slab_energy

get_energy(loaded_reactions[5]['reactionSystems']['star'], 'O2' )

from pprint import pprint
for i in range(1):
    pprint(loaded_reactions[i])

list_adsorbate = ['H', 'O', 'C', 'CO', 'OH', 'N2', 'CO2', 'NH3', 'CH4'] ## 'H2S' 

@ray.remote
def process_reaction(react, list_adsorbate):
    try:
        d_prod = json.loads(react['products'])
        if len(d_prod.keys()) != 1:
            return None

        sc = react['surfaceComposition']
        facet = react['facet']
        star = react['reactionSystems']['star']
        list_sdesc = []
        ase_slab = star
        for ads_symbol in list_adsorbate:
            sdesc = get_energy(ase_slab, ads_symbol)
            list_sdesc.append(sdesc)

        pkey = list(d_prod.keys())[0].replace('star', '')
        pval = d_prod[list(d_prod.keys())[0]]
        pdesc = []

        re = react['reactionEnergy']
        nre = re/pval
        
        eqn = react['Equation']
        ae = react['activationEnergy']
        cc = react['chemicalComposition']
        d_cvr = json.loads(react['coverages'])
        dft_code = react['dftCode']
        dft_func = react['dftFunctional']
        pubid = react['pubId']
        d_reactants = json.loads(react['reactants'])
        sites = json.loads(react['sites'])
        username = react['username']

        rinfo = {
            'sc': sc,
            'facet': facet,
            'star': str(star.symbols),
            'sdesc': list_sdesc,
            
            'pkey': pkey,
            'pval': pval,
            'pdesc': pdesc,
            
            're': re,
            'nre': nre,
            
            'd_prod': d_prod,
            'eqn': eqn,
            'ae': ae, 
            'cc': cc, 
            'd_cvr': d_cvr, 
            'dft_code': dft_code,
            'dft_func': dft_func,
            'pubid': pubid,
            'd_reactants': d_reactants,
            'sites': sites,
            'username': username
            
        }
        return rinfo
    except Exception as e:
        return None
    
futures = [process_reaction.remote(react, list_adsorbate) for react in loaded_reactions[:]]
results = ray.get(futures)

# Filter out None values
list_rinfo = [r for r in results if r is not None]

print(len(list_rinfo))
print(list_rinfo[0])

# with open('datasets/reactions_info.pickle', 'wb') as f:
#     pickle.dump(list_rinfo, f)

with open('datasets/reactions_info.pickle', 'rb') as f:
    loaded_list_rinfo = pickle.load(f)

print(len(loaded_list_rinfo))
print(loaded_list_rinfo[0])

# for d in loaded_list_rinfo[4000:4001]:
#     pprint(d)

## todo: generate product descriptions

list_pkey =  []
for rinfo in loaded_list_rinfo:
    pkey = rinfo['pkey']
    list_pkey.append(pkey)

list(set(list_pkey))

d_pkey_vs_smiles = {
    ## pubchem: one entry/ multi entry/ previous datasets
    'S': '[S]',
    'O': '[O]',
    'CH': '[CH]', 
    'H': '[H]',
    'SH': '[SH]',    
    'NH': '[NH]',
    'CH3': '[CH3]',   
    'H2O': 'O',
    'C': '[C]',   
    'N': '[N]',    
    'OH': '[OH]',
    'CH2': '[CH2]', 
}

## pretrained model
tokenizer = AutoTokenizer.from_pretrained("mrm8488/chEMBL26_smiles_v2")
model = AutoModel.from_pretrained("mrm8488/chEMBL26_smiles_v2")
fe = pipeline('feature-extraction', model=model, tokenizer=tokenizer, device=-1) ## device= (0 for GPU, -1 for CPU)

list_pkey= list(d_pkey_vs_smiles.keys())
list_smiles= list(d_pkey_vs_smiles.values())
list_emb_1 = fe(list_smiles)

list_emb_2 = []
for emb_1 in list_emb_1:
    emb_2 = np.mean(np.array(emb_1[0]), axis=0)
    list_emb_2.append(emb_2)
arr_emb = np.array(list_emb_2)
print(arr_emb.shape)

d_pkey_vs_pdesc = {}
for i, pkey in enumerate(list_pkey):
    pdesc = arr_emb[i]
    d_pkey_vs_pdesc[pkey] = pdesc    

for irinfo, rinfo in enumerate(loaded_list_rinfo):
    pkey = rinfo['pkey']
    pdesc = list(d_pkey_vs_pdesc[pkey])
    rinfo['pdesc'] = pdesc
    if irinfo%10000==0:
        print(irinfo)

# Now, to convert this into a DataFrame:
df = pd.DataFrame(loaded_list_rinfo)

df.head(2)

print(df.shape)

df = df.sort_values(by='nre')
df = df.groupby(['star', 'pkey']).first().reset_index()
print(df.shape)

list_scols = ['s' + str(i) for i in range(len(list_adsorbate))]

df[list_scols] = pd.DataFrame(df.sdesc.tolist(), index= df.index)

print(df.shape)

list_pcols = ['p' + str(i) for i in range(int(arr_emb[0].shape[0]))]

try:
    df[list_pcols] = pd.DataFrame(df.pdesc.tolist(), index= df.index)
except Exception as e:
    print(e)

print(df.shape)

df.head(2)

# df.to_pickle('datasets/reactions_info_df.pickle')

df = pd.read_pickle('datasets/reactions_info_df.pickle')

print(df.shape)

df = df[list_scols + list_pcols + ['nre']]

df.shape

df.head(2)

# df.to_pickle('datasets/reactions_info_df_dataset.pickle')

df = pd.read_pickle('datasets/reactions_info_df_dataset.pickle')

df.shape
