==> exp_IMR_multitaskLearner_multi_surface/script_7d.ipynb <==
## comment

# explore the data from OCP

## import

import pandas as pd
import numpy as np
import random
import ray
from itertools import combinations
import torch.nn as nn
import torch
import math
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge, ElasticNet
from sklearn.kernel_ridge import KernelRidge
from sklearn.svm import SVR
from sklearn.metrics import mean_squared_error, mean_absolute_error
import matplotlib.pyplot as plt
from sklearn.model_selection import GridSearchCV
import warnings
warnings.filterwarnings('ignore')
from sklearn.decomposition import PCA
import torch.optim as optim
from torch.utils.data import DataLoader, TensorDataset
import sys
import os
import logging
from sklearn.model_selection import GroupShuffleSplit
from sklearn.pipeline import Pipeline
import gc
from pprint import pprint

## variables 

curdir = ''
logging.basicConfig(
    filename=curdir + 'logger.log', 
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s'
)
trial_no = 0
seed = 42 + trial_no
np.random.seed(seed)
random.seed(seed)
torch.manual_seed(seed)
if torch.cuda.is_available():
    torch.cuda.manual_seed_all(seed)

####################################
## ocp
####################################

df_ocp1 = pd.read_parquet(curdir + 'datasets/ocp_reactions_info_df.parquet')

df_ocp1.head(2)

def convert_to_int(x):
    try:
        return str(''.join(map(str, x)))
    except ValueError:
        print(f"Error with value: {x}")
        logging.info(f"Error with value: {x}")
        return x

df_ocp1.rename(columns={'energy': 'nre'}, inplace=True)
df_ocp1['miller_index'] = df_ocp1['miller_index'].apply(convert_to_int)
list_scols = ['s' + str(i) for i in range(9)]
list_pcols = ['p' + str(i) for i in range(768)]
list_cols = ['anomaly', 'bulk_mpid', 'miller_index'] + list_scols + list_pcols + ['nre']

list_pkey = sorted(list(df_ocp1['pkey'].unique()))

# list_pkey

df_ocp1 = df_ocp1[list_cols]

print(df_ocp1.shape)

print(df_ocp1.columns)

# 1. Data Preparation:

cases = {
    'case_1': [0],
    'case_2': [0,1],
    'case_3': [0,1,2],
    'case_4': [0,1,2,3],
    'case_5': df_ocp1['anomaly'].unique()
}
sample_counts = {}
filtered_dfs = {}
for case, values in cases.items():
    filtered_df = df_ocp1[df_ocp1['anomaly'].isin(values)]
    filtered_dfs[case] = filtered_df
    sample_counts[case] = len(filtered_df)
print(sample_counts)

cases = {
    'case_1': [0],
    'case_2': [0,1],
    'case_3': [0,1,2],
    'case_4': [0,1,2,3],
    'case_5': df_ocp1['anomaly'].unique()
}

sample_counts = {}
filtered_dfs = {}
for case, values in cases.items():
    filtered_df = df_ocp1[df_ocp1['anomaly'].isin(values)]
    
    # Assuming 'anomaly' column contains the values for which you want to detect outliers
    mean_val = filtered_df['nre'].mean()
    std_val = filtered_df['nre'].std()

    # Filter rows
    print(mean_val, std_val)
    no_outliers_df = filtered_df[
        (filtered_df['nre'] >= mean_val - 1*std_val) &
        (filtered_df['nre'] <= mean_val + 1*std_val)
    ]
    
    filtered_dfs[case] = no_outliers_df
    sample_counts[case] = len(no_outliers_df)

print(sample_counts)

# 2. Preprocess Data:

# Function to preprocess data
def preprocess_data(train, test, use_pca=False):
    features = list(train.columns[3:-1])  # s0 to p767
    X_train = train[features]
    X_test = test[features]
    y_train = train['nre']
    y_test = test['nre']
    
    scaler = StandardScaler()
    # X_train_scaled = scaler.fit_transform(X_train)
    # X_test_scaled = scaler.transform(X_test)
    X_train_scaled = X_train
    X_test_scaled = X_test    

    if use_pca:
        pca = PCA(0.9)
        X_train_scaled = pca.fit_transform(X_train_scaled)
        X_test_scaled = pca.transform(X_test_scaled)
    return X_train_scaled, X_test_scaled, y_train, y_test

# 3. Data Splitting based on groups:

# Define a function to split based on groups
def split_data(df, group_columns):
    gss = GroupShuffleSplit(n_splits=1, train_size=0.75, random_state=seed)
    for train_idx, test_idx in gss.split(df, groups=df[group_columns].astype(str).apply(lambda x: '_'.join(x), axis=1)):
        train = df.iloc[train_idx]
        test = df.iloc[test_idx]
    return train, test

train_test_splits = {}
for case, filtered_df in filtered_dfs.items():
    train, test = split_data(filtered_df, ['bulk_mpid', 'miller_index'])
    train_test_splits[case] = (train, test)
    
    # Calculate unique groups for total, train, and test datasets
    total_groups = filtered_df[['bulk_mpid', 'miller_index']].drop_duplicates().shape[0]
    train_groups = train[['bulk_mpid', 'miller_index']].drop_duplicates().shape[0]
    test_groups = test[['bulk_mpid', 'miller_index']].drop_duplicates().shape[0]

    # Print the results
    print(f"Case: {case}")
    print(f"Shape: {train.shape}, {test.shape}")    
    print(f"Total groups: {total_groups}")
    print(f"Training groups: {train_groups}")
    print(f"Testing groups: {test_groups}\n")    

# 4. Model Training:

results = {}

# Models & hyperparameters
models = {
    'Ridge Regression': {
        'model': Ridge(),
        'params': {
            'alpha': [1.0],
        }
    },
    'Elastic Regression': {
        'model': ElasticNet(),
        'params': {
            'alpha': [1.0],
            'l1_ratio': [0.5]
        }
    },
    # 'Kernel Ridge Regression': {
    #     'model': KernelRidge(),
    #     'params': {
    #         'alpha': [1],
    #         'kernel': ['linear'],
    #         'degree': [3]
    #     }
    # },
    # 'Support Vector Regression': {
    #     'model': SVR(),
    #     'params': {
    #         'C': [1.0],
    #         'kernel': ['rbf'],
    #         'gamma': ['scale']
    #     }
    # }
}

# Create a function to plot and save the figures
def plot_and_save(predicted, true, case, model_name, transformation, folder='outputs'):
    plt.figure(figsize=(5, 3))
    plt.scatter(true, predicted, alpha=0.5)
    plt.xlabel('True')
    plt.ylabel('Predicted')
    plt.title(f"{case}: {model_name} ({transformation})")
    plt.grid(True)
    
    # Save the figure
    # Ensure the 'outputs' directory exists
    if not os.path.exists(curdir + folder):
        os.makedirs(curdir + folder)
    
    plt.savefig(f"{curdir}{folder}/{case}D_{model_name}_{transformation}.png")
    plt.close()

# Train and evaluate models
selected_cases = ['case_1', 'case_2', 'case_3', 'case_4', 'case_5'] 
for case, (train, test) in train_test_splits.items():
    if case not in selected_cases:  # Skip the cases not in selected_cases
        continue    
    print(f"case: {case}")
    
    #############################################################################################################################
    X_train, X_test, y_train, y_test = preprocess_data(train, test, use_pca=False) # For non-PCA version
    print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)
    for model_name, model_details in models.items():
        print(model_name)
        clf = GridSearchCV(model_details['model'], model_details['params'], cv=5, scoring='neg_mean_absolute_error', n_jobs=1)
        clf.fit(X_train, y_train)    
        predicted = clf.predict(X_test)
        mae = -clf.score(X_test, y_test)
        params = clf.best_params_
        mean_target = y_test.abs().mean()
        # Store the results
        results[(case, model_name, 'non_pca')] = (mae, mean_target, params)
        
        # Plot
        plot_and_save(predicted, y_test, case, model_name, 'non_pca')
        del predicted
    del X_train, X_test, y_train, y_test
    gc.collect()
    
    #############################################################################################################################        
    X_train, X_test, y_train, y_test = preprocess_data(train, test, use_pca=True) # For PCA version
    print(X_train.shape, X_test.shape, y_train.shape, y_test.shape)    
    for model_name, model_details in models.items():
        print(model_name)
        clf = GridSearchCV(model_details['model'], model_details['params'], cv=5, scoring='neg_mean_absolute_error', n_jobs=1)
        clf.fit(X_train, y_train)
        predicted = clf.predict(X_test)
        mae = -clf.score(X_test, y_test)
        params = clf.best_params_
        mean_target = y_test.abs().mean()
        # Store the results
        results[(case, model_name, 'pca')] = (mae, mean_target, params)
        
        # Plot
        plot_and_save(predicted, y_test, case, model_name, 'pca')
        del predicted
        
    del X_train, X_test, y_train, y_test
    gc.collect()

pprint(results)
