==> exp_IMR_multitaskLearner_multi_surface/script_9_cathub.ipynb <==
## for the follwing code
#         # Embedding block.
#         x = self.emb(data.atomic_numbers.long(), rbf, i, j)
#         P = self.output_blocks[0](x, rbf, i, num_nodes=pos.size(0))
#         print(x.shape, P.shape)
        
#         ## change
#         def capture_embeddings(module, input, output):
#             self.embs.append(output.detach())  # Store the output embeddings
            
#         ## change
#         hook_handle = self.interaction_blocks[-1].register_forward_hook(capture_embeddings)
            
#         # Interaction blocks.
#         for interaction_block, output_block in zip(
#             self.interaction_blocks, self.output_blocks[1:]
#         ):
#             x = interaction_block(x, rbf, sbf, idx_kj, idx_ji)
#             P += output_block(x, rbf, i, num_nodes=pos.size(0))
#             print(x.shape, P.shape)

## getting following output
# Atoms(symbols='H5Au48Co8O17', pbc=True, cell=[5.878883441, 20.365049623, 20.81515], calculator=OCPCalculator(...))
# torch.Size([3477, 256]) torch.Size([78, 1])
# torch.Size([3477, 256]) torch.Size([78, 1])
# torch.Size([3477, 256]) torch.Size([78, 1])
# torch.Size([3477, 256]) torch.Size([78, 1])
# Atoms(symbols='CH9Au48Co8O17', pbc=True, cell=[5.878883441, 20.365049623, 20.81515], calculator=OCPCalculator(...))
# torch.Size([3533, 256]) torch.Size([83, 1])
# torch.Size([3533, 256]) torch.Size([83, 1])
# torch.Size([3533, 256]) torch.Size([83, 1])
# torch.Size([3533, 256]) torch.Size([83, 1])
# -1.2942085266113281

import os
import numpy as np
import ase.io
from ase.constraints import FixAtoms
from ase.build import add_adsorbate, molecule, surface
from pymatgen.ext.matproj import MPRester
from pymatgen.core.surface import generate_all_slabs, SlabGenerator
from pymatgen.io.ase import AseAtomsAdaptor
from ocpmodels.common.relaxation.ase_utils import OCPCalculator
import torch
import pickle
import json
import pandas as pd

config_yml_path = "ocp/configs/is2re/all/dimenet_plus_plus/dpp.yml"
checkpoint_path = "dimenetpp_all.pt"
# Define the calculator
calc = OCPCalculator(config_yml=config_yml_path, checkpoint=checkpoint_path)

calc

with open('datasets/reactions.pickle', 'rb') as f:
    loaded_reactions = pickle.load(f)

len(loaded_reactions)

# list_rinfo = []
# for ir, r in enumerate(loaded_reactions):
#     try:
#         if ir%1000==0:
#             print(ir, len(list_rinfo))
    
#         ## retrieve slab and adsorbate
#         ase_slab = r['reactionSystems']['star']
#         d_prod = json.loads(r['products'])
#         if len(d_prod.keys()) != 1:
#             continue
#         pkey = list(d_prod.keys())[0].replace('star', '')
#         ads_symbol = pkey    

#         ## get slab energy and embeddings
#         adslab = ase_slab.copy()
#         adslab.calc = calc
#         # print(adslab)
#         slab_energy = adslab.get_potential_energy()
#         slab_embs = torch.mean(calc.trainer.model.module.embs[0], dim=0)    

#         ## get slab + adsorbate energy and embeddings
#         # add the molecule
#         adsorbate = molecule(ads_symbol)
#         add_adsorbate(adslab, adsorbate, 3, offset=(1, 1))
#         # print(adslab)
#         ads_energy = adslab.get_potential_energy()
#         ads_embs = torch.mean(calc.trainer.model.module.embs[0], dim=0)    
        
#         ## additional info
#         sc = r['surfaceComposition']
#         facet = r['facet']
#         pval = d_prod[list(d_prod.keys())[0]]        
#         re = r['reactionEnergy']
#         nre = re/pval
        
#         eqn = r['Equation']
#         ae = r['activationEnergy']
#         cc = r['chemicalComposition']
#         d_cvr = json.loads(r['coverages'])
#         dft_code = r['dftCode']
#         dft_func = r['dftFunctional']
#         pubid = r['pubId']
#         d_reactants = json.loads(r['reactants'])
#         sites = json.loads(r['sites'])
#         username = r['username']

#         rinfo = {
#             'ase_slab': ase_slab,            
#             'd_prod': d_prod,            
#             'pkey': pkey,
#             'slab_embs': slab_embs,
#             'ads_embs': ads_embs,
            
#             'sc': sc,
#             'facet': facet,            
#             'pval': pval,            
#             're': re,
#             'nre': nre,
            
#             'eqn': eqn,
#             'ae': ae, 
#             'cc': cc, 
#             'd_cvr': d_cvr, 
#             'dft_code': dft_code,
#             'dft_func': dft_func,
#             'pubid': pubid,
#             'd_reactants': d_reactants,
#             'sites': sites,
#             'username': username            
#         }
#         list_rinfo.append(rinfo)
#     except Exception as e:
#         pass
    

# with open('datasets/reactions_info_2.pickle', 'wb') as f:
#     pickle.dump(list_rinfo, f)

with open('datasets/reactions_info_2.pickle', 'rb') as f:
    loaded_list_rinfo = pickle.load(f)

len(loaded_list_rinfo)

# Convert tensor values in 'ads_embs' to lists of floats
for item in loaded_list_rinfo:
    item['ads_embs'] = [float(x) for x in item['ads_embs']]

# Convert list of dictionaries to DataFrame
df = pd.DataFrame(loaded_list_rinfo)

# Split 'ads_embs' into separate columns
ads_embs_cols = df['ads_embs'].apply(pd.Series)
ads_embs_cols.columns = ['e' + str(i) for i in range(256)]

# Drop the original 'ads_embs' column and concatenate the new columns
df = df.drop('ads_embs', axis=1)
df = pd.concat([df, ads_embs_cols], axis=1)

df = df.sort_values(by='nre')
df = df.groupby(['sc', 'facet', 'pkey']).first().reset_index()
print(df.shape)

df = df[['e' + str(i) for i in range(256)] + ['nre']]

df.shape

df.head(2)

# df.to_pickle('datasets/reactions_info_df_2.pickle')

df = pd.read_pickle('datasets/reactions_info_df_2.pickle')

df.shape

df.head()
