==> exp_IMR_multitaskLearner_multi_surface/script_10_random_data_check.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

criteria = [
    {'symbol': 'Pt', 'miller_index': [1, 1, 1]},
    {'symbol': 'Ru', 'miller_index': [[0, 0, 0, 1], [0, 0, 1]]},
    {'symbol': 'Ni', 'miller_index': [1, 1, 1]},
    {'symbol': 'Co', 'miller_index': [[0, 0, 0, 1], [0, 0, 1]]},
    {'symbol': 'Fe', 'miller_index': [1, 1, 0]},
    {'symbol': 'Cu', 'miller_index': [1, 1, 1]}
]

df = pd.read_parquet('datasets/ocp_reactions_info_df.parquet')

df2 = df[['ads_symbols', 'bulk_symbols', 'miller_index', 'energy']]

df2[
    (df2['bulk_symbols'].isin(['Pt', 'Ru', 'Ni', 'Co', 'Fe', 'Cu'])) &
    (df2['ads_symbols'] == '*C')
].head()


merged_df = pd.DataFrame()

a = df2[
    (df2['bulk_symbols']=='Pt') &
    (df2['ads_symbols'] == '*C')
]

b= df2[
    (df2['bulk_symbols']=='Ru') &
    (df2['ads_symbols'] == '*C')
]

c=df2[
    (df2['bulk_symbols']=='Ni') &
    (df2['ads_symbols'] == '*C')
]

d=df2[
    (df2['bulk_symbols']=='Co') &
    (df2['ads_symbols'] == '*C')
]

e=df2[
    (df2['bulk_symbols']=='Fe') &
    (df2['ads_symbols'] == '*C')
]

f=df2[
    (df2['bulk_symbols']=='Cu') &
    (df2['ads_symbols'] == '*C')
]

a.shape, b.shape, c.shape, d.shape, e.shape, f.shape

merged_df = pd.concat([a, b, c, d, e, f])
merged_df

import pandas as pd

# Assuming df2 is your DataFrame and it's already defined in your environment
# df2 = ...

# List of bulk symbols to filter
bulk_symbols_list = ['Pt', 'Ru', 'Ni', 'Co', 'Fe', 'Cu']

# Initialize an empty DataFrame for the merged results
merged_df = pd.DataFrame()

# Iterate through the bulk symbols, filter, and merge the results
for symbol in bulk_symbols_list:
    temp_df = df2[
        (df2['bulk_symbols'] == symbol) &
        (df2['ads_symbols'] == '*C')
    ].head(1)
    merged_df = pd.concat([merged_df, temp_df])

# Display the merged DataFrame
print(merged_df.head(len(bulk_symbols_list)))

filtered_df = pd.DataFrame()
for crit in criteria:
    symbol = crit['symbol']
    miller_indices = crit['miller_index']
    if not isinstance(miller_indices[0], list):
        miller_indices = [miller_indices]

    for index in miller_indices:
        temp_df = df2[(df2['bulk_symbols']==symbol) & 
                     (df2['miller_index'].apply(lambda x: all([a == b for a, b in zip(x, index)]))) & 
                     (df2['ads_symbols'] == '*C')]
        temp_df = temp_df.sort_values(by='bulk_symbols').head(1)        
        filtered_df = pd.concat([filtered_df, temp_df])
        print(crit, index, temp_df)


filtered_df

import pickle

with open('datasets/reactions_info.pickle', 'rb') as f:
    loaded_list_rinfo = pickle.load(f)

dfc1 = pd.DataFrame(loaded_list_rinfo)

dfc2 = dfc1[['pkey', 'sc', 'facet', 'nre']]

dfc2[
    (dfc2['sc']=='Pt') &
    (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head(1)

dfc2[
    (dfc2['sc']=='Ru') &
    # (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head()

dfc2[
    (dfc2['sc']=='Ni') &
    # (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head()

dfc2[
    (dfc2['sc']=='Co') &
    # (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head()

dfc2[
    (dfc2['sc']=='Fe') &
    # (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head()

dfc2[
    (dfc2['sc']=='Cu') &
    # (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head()

dfc2[
    (dfc2['sc']=='Pt') &
    (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head(1)

dfc2[
    (dfc2['sc']=='Cu') &
    (dfc2['facet'] == '111') &
    (dfc2['pkey'] == 'C')
].head(1)

conditions = [
    {'sc': 'Pt', 'facet': '111', 'pkey': 'C'},
    {'sc': 'Cu', 'facet': '111', 'pkey': 'C'}
]
merged_df = pd.DataFrame()
for condition in conditions:
    temp_df = dfc2[
        (dfc2['sc'] == condition['sc']) &
        (dfc2['facet'] == condition['facet']) &
        (dfc2['pkey'] == condition['pkey'])
    ].head(1)
    
    merged_df = pd.concat([merged_df, temp_df])
print(merged_df.head())
